requirements.txt
django
django-rest-framework
pdfplumber
mistralai
python-dotenv